requirements.txt
speedtest-cli
tqdm
python-dotenv
openai
asciichartpy
requests